@@ -6,39 +6,41 @@
     "minimum-stability": "dev",
     "prefer-stable": true,
     "require": {
-        "php": "^7.4 | ^8.0",
+        "php": "^8.1",
         "ext-ctype": "*",
         "ext-iconv": "*",
         "ext-sqlite3": "*",
-        "box/spout": "^3.1",
         "composer/package-versions-deprecated": "^1.11",
         "doctrine/annotations": "^1.11",
         "doctrine/doctrine-bundle": "^2.2",
         "doctrine/doctrine-migrations-bundle": "^3.0",
         "doctrine/orm": "^2.8",
-        "guzzlehttp/guzzle": "^7.2",
         "neitanod/forceutf8": "^2.0",
-        "symfony/console": "5.2.*",
-        "symfony/dotenv": "5.2.*",
-        "symfony/filesystem": "5.2.*",
-        "symfony/finder": "5.2.*",
-        "symfony/flex": "^1.3.1",
-        "symfony/framework-bundle": "5.2.*",
+        "symfony/asset": "6.1.*",
+        "symfony/console": "^6.1",
+        "symfony/dotenv": "^6.1",
+        "symfony/filesystem": "^6.1",
+        "symfony/finder": "^6.1",
+        "symfony/flex": "^2.2",
+        "symfony/framework-bundle": "^6.1",
+        "symfony/http-client": "^6.1",
         "symfony/monolog-bundle": "^3.6",
-        "symfony/proxy-manager-bridge": "5.2.*",
-        "symfony/twig-bundle": "5.2.*",
-        "symfony/yaml": "5.2.*",
+        "symfony/proxy-manager-bridge": "^6.1",
+        "symfony/runtime": "6.1.*",
+        "symfony/twig-bundle": "^6.1",
+        "symfony/yaml": "^6.1",
         "twig/twig": "^3.0"
     },
     "require-dev": {
-        "escapestudios/symfony2-coding-standard": "^3.11",
-        "friendsofphp/php-cs-fixer": "^2.17",
-        "phpunit/phpunit": "^9.5",
-        "psalm/plugin-symfony": "^2.1",
-        "symfony/browser-kit": "^5.2",
-        "symfony/css-selector": "^5.2",
+        "escapestudios/symfony2-coding-standard": "^3.12",
+        "friendsofphp/php-cs-fixer": "^3.8",
+        "kubawerlos/php-cs-fixer-custom-fixers": "^3.11",
+        "psalm/plugin-symfony": "^3.1",
+        "rector/rector": "^0.13.7",
+        "symfony/browser-kit": "^6.1",
+        "symfony/css-selector": "^6.1",
         "symfony/maker-bundle": "^1.26",
-        "symfony/phpunit-bridge": "^5.3",
+        "symfony/phpunit-bridge": "^6.1",
         "vimeo/psalm": "^4.3"
     },
     "config": {
@@ -48,7 +50,8 @@
         },
         "sort-packages": true,
         "allow-plugins": {
-            "symfony/flex": true
+            "symfony/flex": true,
+            "symfony/runtime": true
         }
     },
     "autoload": {
@@ -80,11 +83,11 @@
         "post-update-cmd": [
             "@auto-scripts"
         ],
-        "check-coding-standards": [
-            "php-cs-fixer --config=.php_cs.dist fix --dry-run --format=checkstyle"
+        "coding-standards-check": [
+            "php-cs-fixer fix --dry-run --format=checkstyle"
         ],
-        "apply-coding-standards": [
-            "php-cs-fixer --config=.php_cs.dist fix"
+        "coding-standards-apply": [
+            "php-cs-fixer fix"
         ],
         "psalm": [
             "./vendor/bin/psalm --show-info=true"
@@ -95,8 +98,8 @@
     },
     "extra": {
         "symfony": {
-            "allow-contrib": false,
-            "require": "5.2.*"
+            "allow-contrib": true,
+            "require": "6.1.*"
         }
     }
 }
