@@ -4,10 +4,7 @@ namespace App\Service;
 
 use App\Entity\SearchFeed;
 use App\Repository\SearchFeedRepository;
-use Box\Spout\Common\Entity\Row;
-use Box\Spout\Reader\Common\Creator\ReaderEntityFactory;
-use Box\Spout\Reader\CSV\Sheet;
-use Box\Spout\Writer\Common\Creator\WriterEntityFactory;
+use Doctrine\DBAL\Exception;
 use Doctrine\ORM\EntityManagerInterface;
 use ForceUTF8\Encoding;
 
@@ -16,25 +13,20 @@ use ForceUTF8\Encoding;
  */
 class ParseSearchFeedService
 {
-    private string $projectDir;
-    private string $destinationDirectory;
-    private SearchFeedRepository $searchFeedRepos;
-    private EntityManagerInterface $em;
-
     /**
      * ParseSearchFeedService constructor.
      *
-     * @param string $bindProjectDir
-     * @param string $bindDestinationDirectory
-     * @param EntityManagerInterface $entityManager
-     * @param SearchFeedRepository $searchFeedRepository
+     * @param string $destinationDirectory
+     * @param EntityManagerInterface $em
+     * @param SearchFeedRepository $searchFeedRepos
+     * @param CsvReaderService $CsvReader
      */
-    public function __construct(string $bindProjectDir, string $bindDestinationDirectory, EntityManagerInterface $entityManager, SearchFeedRepository $searchFeedRepository)
-    {
-        $this->projectDir = $bindProjectDir;
-        $this->destinationDirectory = $bindDestinationDirectory;
-        $this->em = $entityManager;
-        $this->searchFeedRepos = $searchFeedRepository;
+    public function __construct(
+        private readonly string $destinationDirectory,
+        private readonly EntityManagerInterface $em,
+        private readonly SearchFeedRepository $searchFeedRepos,
+        private readonly CsvReaderService $CsvReader
+    ) {
     }
 
     /**
@@ -45,85 +37,68 @@ class ParseSearchFeedService
      * @param string $filename
      *   If provided the file will be used as input else file will be downloaded
      *
-     * @return \Generator
      *   Yield for every 500 rows
-     *
-     * @throws \Box\Spout\Common\Exception\IOException
-     * @throws \Box\Spout\Reader\Exception\ReaderNotOpenedException
      */
     public function parse(string $filename): \Generator
     {
-        $reader = ReaderEntityFactory::createCSVReader();
-        $reader->open($filename);
-
         $rowsCount = 0;
         $rowsInserted = 0;
+        $rowsUpdated = 0;
 
         // Bookkeeping between batches.
         $entities = [];
 
-        $this->em->getConnection()->beginTransaction();
+        $iterator = $this->CsvReader->read($filename);
+        foreach ($iterator as $line) {
+            $searchYear = (int) $line[0];
+            $searchWeek = (int) $line[1];
 
-        /* @var Sheet $sheet */
-        foreach ($reader->getSheetIterator() as $sheet) {
-            /* @var Row $row */
-            foreach ($sheet->getRowIterator() as $row) {
-                $searchYear = (int) $row->getCellAtIndex(0)->getValue();
-                $searchWeek = (int) $row->getCellAtIndex(1)->getValue();
+            ++$rowsCount;
 
-                ++$rowsCount;
+            // Yield progress.
+            if (0 == $rowsCount % 500) {
+                yield ['processed' => $rowsCount, 'inserted' => $rowsInserted, 'updated' => $rowsUpdated];
+            }
 
-                // Yield progress.
-                if (0 == $rowsCount % 500) {
-                    yield ['processed' => $rowsCount, 'inserted' => $rowsInserted];
-                }
+            if ($this->isFromPeriod($searchYear, $searchWeek)) {
+                $searchKey = htmlspecialchars_decode((string) $line[2]);
+                $search_count = (int) $line[3];
 
-                if ($this->isFromPeriod($searchYear, $searchWeek)) {
-                    $searchKey = $row->getCellAtIndex(2)->getValue();
-                    $search_count = (int) $row->getCellAtIndex(3)->getValue();
+                // We exclude complex search strings.
+                if ($this->isValid($searchKey)) {
+                    $entities[$searchKey] = array_key_exists($searchKey, $entities) ? $entities[$searchKey] : $this->searchFeedRepos->findOneBy(['search' => $searchKey]);
+                    if (is_null($entities[$searchKey])) {
+                        $entities[$searchKey] = new SearchFeed();
+                        $entities[$searchKey]->setYear($searchYear);
+                        $entities[$searchKey]->setWeek($searchWeek);
+                        $entities[$searchKey]->setSearch($searchKey);
 
-                    // We exclude complex search strings.
-                    if ($this->isValid($searchKey)) {
+                        $this->em->persist($entities[$searchKey]);
                         ++$rowsInserted;
+                    } else {
+                        ++$rowsUpdated;
+                    }
+                    $entities[$searchKey]->incriminateLongPeriod($search_count);
 
-                        $entities[$searchKey] = array_key_exists($searchKey, $entities) ? $entities[$searchKey] : $this->searchFeedRepos->findOneBy(['search' => $searchKey]);
-                        if (is_null($entities[$searchKey])) {
-                            $entities[$searchKey] = new SearchFeed();
-                            $entities[$searchKey]->setYear($searchYear);
-                            $entities[$searchKey]->setWeek($searchWeek);
-                            $entities[$searchKey]->setSearch($searchKey);
+                    if ($this->isFromPeriod($searchYear, $searchWeek, 4)) {
+                        $entities[$searchKey]->incriminateShortPeriod($search_count);
+                    }
 
-                            $this->em->persist($entities[$searchKey]);
-                        }
-                        $entities[$searchKey]->incriminateLongPeriod($search_count);
+                    // Make it stick for every 5000 entities loaded into memory.
+                    if (0 === count($entities) % 5000) {
+                        $this->em->flush();
+                        $this->em->clear();
 
-                        if ($this->isFromPeriod($searchYear, $searchWeek, 4)) {
-                            $entities[$searchKey]->incriminateShortPeriod($search_count);
-                        }
-
-                        // Make it stick for every 5000 rows.
-                        if (0 === $rowsInserted % 5000) {
-                            $this->em->flush();
-                            $this->em->getConnection()->commit();
-                            $this->em->clear();
-
-                            $entities = [];
-                            gc_collect_cycles();
-
-                            // Start new transaction for the next batch.
-                            $this->em->getConnection()->beginTransaction();
-                        }
+                        $entities = [];
+                        gc_collect_cycles();
                     }
                 }
             }
-
-            // Make it stick.
-            $this->em->flush();
-            $this->em->getConnection()->commit();
-            $this->em->clear();
         }
 
-        $reader->close();
+        // Make it stick.
+        $this->em->flush();
+        $this->em->clear();
     }
 
     /**
@@ -133,14 +108,10 @@ class ParseSearchFeedService
      *   Number of rows to output
      * @param string $filename
      *   The filename to write to in the public folder
-     *
-     * @throws \Box\Spout\Common\Exception\IOException
-     * @throws \Box\Spout\Writer\Exception\WriterNotOpenedException
      */
     public function writeFile(int $rows = 5000, string $filename = 'searchdata.csv'): void
     {
-        $writer = WriterEntityFactory::createCSVWriter();
-        $writer->openToFile($this->destinationDirectory.'/'.$filename);
+        $file = fopen($this->destinationDirectory.'/'.$filename, 'w');
 
         $query = $this->em->createQueryBuilder()
             ->select('s')
@@ -155,17 +126,16 @@ class ParseSearchFeedService
             $search = Encoding::toUTF8($entity->getSearch());
 
             $values = [$search, $entity->getLongPeriod(), $entity->getShortPeriod()];
-            $rowFromValues = WriterEntityFactory::createRowFromArray($values);
-            $writer->addRow($rowFromValues);
+            fputcsv($file, $values);
         }
 
-        $writer->close();
+        fclose($file);
     }
 
     /**
      * Reset the database table (truncate it).
      *
-     * @throws \Doctrine\DBAL\Exception
+     * @throws Exception
      */
     public function reset(): void
     {
@@ -181,12 +151,9 @@ class ParseSearchFeedService
      *   Week number
      * @param int $period
      *   Weeks from now that the year and week should be with in
-     *
-     * @return bool
      */
     private function isFromPeriod(int $year, int $week, int $period = 52): bool
     {
-        // @TODO: Change this to use timestamps.
         $date = new \DateTime();
         $nowYear = (int) $date->format('Y');
         $nowWeek = (int) $date->format('W');
@@ -210,12 +177,11 @@ class ParseSearchFeedService
      * @param string $key
      *   The search key
      *
-     * @return bool
      *   The result of the validation
      */
     private function isValid(string $key): bool
     {
-        if (false !== strpos($key, '=') || false !== strpos($key, '(') || false !== strpos($key, '*')) {
+        if (str_contains($key, '=') || str_contains($key, '(') || str_contains($key, '*')) {
             return false;
         }
 
